Add files via upload
ABC-PMC in Python for running initial analyses

note: truepos file contains all the (x_true,y_true) from the original data. It is used for getting the 'TrueData'. in the code called 'TrueModel'. That's because the FM used in this step is not the one used for getting the data in the original table.